remove redundant code and streamline Flask app initialization.
