+++ backend/app/__init__.py
@@ -1,1 +1,5 @@
 # This file is intentionally empty to make this directory a Python package 
+
+from app.core.factory import create_app
+
+__all__ = ['create_app']
