+++ backend/app/api/models/__init__.py
@@ -1,0 +1,1 @@
+# Models package for data models
